+++ serialization/src/test/java/type/FixedTest.java
@@ -17,7 +17,6 @@ public class FixedTest {
     UInt64 sinceUInt64 = new UInt64(1L);
     Struct inputs = new Struct(sinceUInt64, new Struct(txHash, index));
     Fixed<Struct> structFixed = new Fixed<>(Collections.singletonList(inputs));
-    System.out.println(Numeric.toHexString(structFixed.toBytes()));
     Assertions.assertArrayEquals(
         Numeric.hexStringToByteArray(
             "0x010000000100000000000000000000000000000000000000000000000000000000000000000000000000000095729694"),